src/api/requirements.txt
fastapi
uvicorn
sagemaker
scikit-learn==1.2.1
python-dotenv